+++ src/services/speakers/speakers-router.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { StatusCodes } from "http-status-codes";
-import { SpeakerValidator } from "./speakers-schema";
+import { SpeakerValidator } from "./speakers-validators";
 import { Database } from "../../database";
 import RoleChecker from "../../middleware/role-checker";
 import { Role } from "../auth/auth-models";
