+++ src/services/speakers/speakers-schema.ts
@@ -1,17 +1,5 @@
 import { Schema } from "mongoose";
-import { z } from "zod";
 import { v4 as uuidv4 } from "uuid";
-
-// Zod schema for speaker
-export const SpeakerValidator = z.object({
-    speakerId: z.coerce.string().default(() => uuidv4()),
-    name: z.string(),
-    title: z.string(),
-    bio: z.string(),
-    eventTitle: z.string(),
-    eventDescription: z.string(),
-    imgUrl: z.string(),
-});
 
 // Mongoose schema for speaker
 export const SpeakerSchema = new Schema({
